+++ src/app/api/images/[contentType]/[id]/route.ts
@@ -1,6 +1,6 @@
 import { forbidden } from "next/navigation";
 import { NextResponse } from "next/server";
-import { getImagesFromStorage } from "@/applications/images/get-images";
+import { getImagesFromStorage } from "@/application-services/images/get-images";
 import { auth } from "@/common/auth/auth";
 
 export const GET = auth(
