+++ src/application-services/contents/delete-contents.ts
