+++ src/application-services/images/delete-images.ts
