+++ src/application-services/news/delete-news.ts
