+++ src/application-services/news/get-news.test.ts
